[function "equalize" independent from join]
Documentation:
- help pages for functions: datetime, equalize, estimateFileSize, tindex

Function:
- started to treat "equalize" as a independent function

Notes:
- changing the way join works currently: join will only work with
similar arrays regarding the dimension, and the user will take care to
make the arrays similar with equalize. The reason for this is that the
user can specify the correct aggregation function, because it is more
transparent which array the target array will be. It will also be easier
to realize the integration of the scidbs "xgrid" function to break down
low resolution arrays into a higher resolution

Fix:
- temporary arrays will now be named with __temp_<array@title>_XXX with
XXX being a random number and the title will be taken from the parent
array